+++ library/common/src/main/java/com/google/android/exoplayer2/util/MimeTypes.java
@@ -123,22 +123,22 @@ public final class MimeTypes {
     customMimeTypes.add(customMimeType);
   }
 
-  /** Returns whether the given string is an audio mime type. */
+  /** Returns whether the given string is an audio MIME type. */
   public static boolean isAudio(@Nullable String mimeType) {
     return BASE_TYPE_AUDIO.equals(getTopLevelType(mimeType));
   }
 
-  /** Returns whether the given string is a video mime type. */
+  /** Returns whether the given string is a video MIME type. */
   public static boolean isVideo(@Nullable String mimeType) {
     return BASE_TYPE_VIDEO.equals(getTopLevelType(mimeType));
   }
 
-  /** Returns whether the given string is a text mime type. */
+  /** Returns whether the given string is a text MIME type. */
   public static boolean isText(@Nullable String mimeType) {
     return BASE_TYPE_TEXT.equals(getTopLevelType(mimeType));
   }
 
-  /** Returns whether the given string is an application mime type. */
+  /** Returns whether the given string is an application MIME type. */
   public static boolean isApplication(@Nullable String mimeType) {
     return BASE_TYPE_APPLICATION.equals(getTopLevelType(mimeType));
   }
@@ -174,13 +174,14 @@ public final class MimeTypes {
    * @param codecs The codecs attribute.
    * @return The derived video mimeType, or null if it could not be derived.
    */
-  public static @Nullable String getVideoMediaMimeType(@Nullable String codecs) {
+  @Nullable
+  public static String getVideoMediaMimeType(@Nullable String codecs) {
     if (codecs == null) {
       return null;
     }
     String[] codecList = Util.splitCodecs(codecs);
     for (String codec : codecList) {
-      String mimeType = getMediaMimeType(codec);
+      @Nullable String mimeType = getMediaMimeType(codec);
       if (mimeType != null && isVideo(mimeType)) {
         return mimeType;
       }
@@ -194,13 +195,14 @@ public final class MimeTypes {
    * @param codecs The codecs attribute.
    * @return The derived audio mimeType, or null if it could not be derived.
    */
-  public static @Nullable String getAudioMediaMimeType(@Nullable String codecs) {
+  @Nullable
+  public static String getAudioMediaMimeType(@Nullable String codecs) {
     if (codecs == null) {
       return null;
     }
     String[] codecList = Util.splitCodecs(codecs);
     for (String codec : codecList) {
-      String mimeType = getMediaMimeType(codec);
+      @Nullable String mimeType = getMediaMimeType(codec);
       if (mimeType != null && isAudio(mimeType)) {
         return mimeType;
       }
@@ -214,7 +216,8 @@ public final class MimeTypes {
    * @param codec The codec identifier to derive.
    * @return The mimeType, or null if it could not be derived.
    */
-  public static @Nullable String getMediaMimeType(@Nullable String codec) {
+  @Nullable
+  public static String getMediaMimeType(@Nullable String codec) {
     if (codec == null) {
       return null;
     }
@@ -235,7 +238,7 @@ public final class MimeTypes {
     } else if (codec.startsWith("vp8") || codec.startsWith("vp08")) {
       return MimeTypes.VIDEO_VP8;
     } else if (codec.startsWith("mp4a")) {
-      String mimeType = null;
+      @Nullable String mimeType = null;
       if (codec.startsWith("mp4a.")) {
         String objectTypeString = codec.substring(5); // remove the 'mp4a.' prefix
         if (objectTypeString.length() >= 2) {
@@ -244,7 +247,7 @@ public final class MimeTypes {
             int objectTypeInt = Integer.parseInt(objectTypeHexString, 16);
             mimeType = getMimeTypeFromMp4ObjectType(objectTypeInt);
           } catch (NumberFormatException ignored) {
-            // ignored
+            // Ignored.
           }
         }
       }
@@ -267,6 +270,10 @@ public final class MimeTypes {
       return MimeTypes.AUDIO_VORBIS;
     } else if (codec.startsWith("flac")) {
       return MimeTypes.AUDIO_FLAC;
+    } else if (codec.startsWith("stpp")) {
+      return MimeTypes.APPLICATION_TTML;
+    } else if (codec.startsWith("wvtt")) {
+      return MimeTypes.TEXT_VTT;
     } else {
       return getCustomMimeTypeForCodec(codec);
     }
@@ -406,7 +413,8 @@ public final class MimeTypes {
    * Returns the top-level type of {@code mimeType}, or null if {@code mimeType} is null or does not
    * contain a forward slash character ({@code '/'}).
    */
-  private static @Nullable String getTopLevelType(@Nullable String mimeType) {
+  @Nullable
+  private static String getTopLevelType(@Nullable String mimeType) {
     if (mimeType == null) {
       return null;
     }
@@ -417,7 +425,8 @@ public final class MimeTypes {
     return mimeType.substring(0, indexOfSlash);
   }
 
-  private static @Nullable String getCustomMimeTypeForCodec(String codec) {
+  @Nullable
+  private static String getCustomMimeTypeForCodec(String codec) {
     int customMimeTypeCount = customMimeTypes.size();
     for (int i = 0; i < customMimeTypeCount; i++) {
       CustomMimeType customMimeType = customMimeTypes.get(i);